Create Generative Adversarial Networks (GANs)
GANs (Generative Adversarial Networks) are considered one of the most modern and fascinating technologies within the field of Deep Learning and Computer Vision. They have gained a lot of attention because they can create fake content. One of the most classic examples is the creation of people who do not exist in the real world to be used to broadcast television programs. This technology is considered a revolution in the field of Artificial Intelligence for producing high quality results, remaining one of the most popular and relevant topics. 

Reference : Prof. Dr. Jones Granatyr